Move makePackageObjPrefixExplicit to TypeUtils
